feat: add quarter-step CRF support & cleanups (#1108)
- Quarter-step CRF will be supported by `x264`, `x265` and `svt-av1`
- Especially `x26*` encoders will heavily benefit from the additional
accuracy / speed
- Runtime detection of SVT-AV1 quarter-step support via test encoding
- Cache quarter-step capability check using `OnceLock`
- Maintain backward compatibility with mainline SVT-AV1
- Remove obsolete SVT-AV1 version check and legacy parsing

+++ av1an-core/src/chunk/mod.rs
@@ -30,7 +30,7 @@ pub struct Chunk {
     // do not break compatibility with output produced by older versions of av1an
     /// Optional target quality CQ level
     #[serde(rename = "per_shot_target_quality_cq")]
-    pub tq_cq:                 Option<u32>,
+    pub tq_cq:                 Option<f32>,
     pub ignore_frame_mismatch: bool,
 }
